enum int8 로 serialize 되는 것 수정

+++ TableGenerateCmd.NetCore/ExportBase.cs
@@ -465,7 +465,7 @@ namespace TableGenerate
                                             var o = enumDescriptor.GetOptions();
                                             var value = o.GetExtension(BitFlagsExtensions.BluePrintType);
                                             column.blue_print_type = value;
-                                            column.primitive_type = BaseType.Int8;
+                                            column.primitive_type = BaseType.Int32;
                                         }
                                         break;
                                     }
